pyproject.toml
[tool.poetry]
name = "gym"
version = "0.1.0"
description = ""
authors = ["sc303030 <lkh6035@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.4"


[tool.poetry.group.dev.dependencies]
django-model-utils = "^4.3.1"
selenium = "^4.7.2"
beautifulsoup4 = "^4.11.1"
requests = "^2.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
